requirements.txt
pydantic==2.10.1
chainlit==1.3.2
groq==0.12.0
python-dotenv==1.0.1
ur-rtde==1.5.9
openai==1.55.3
fire==0.7.0
tqdm==4.67.1
rouge-score==0.1.2
flask==3.1.0
transformers==4.46.3
google-generativeai==0.8.3
# fix temp
httpx==0.27.2
